Abstract common test code into BaseNTestData
