+++ src/test/java/org/apache/commons/codec/binary/Base32OutputStreamTest.java
@@ -117,7 +117,7 @@ public class Base32OutputStreamTest {
         // test random data of sizes 0 thru 150
         final BaseNCodec codec = new Base32();
         for (int i = 0; i <= 150; i++) {
-            final byte[][] randomData = Base32TestData.randomData(codec, i);
+            final byte[][] randomData = BaseNTestData.randomData(codec, i);
             encoded = randomData[1];
             decoded = randomData[0];
             testByChunk(encoded, decoded, 0, LF);
@@ -152,7 +152,7 @@ public class Base32OutputStreamTest {
         // test random data of sizes 0 thru 150
         final BaseNCodec codec = new Base32();
         for (int i = 0; i <= 150; i++) {
-            final byte[][] randomData = Base32TestData.randomData(codec, i);
+            final byte[][] randomData = BaseNTestData.randomData(codec, i);
             encoded = randomData[1];
             decoded = randomData[0];
             testByteByByte(encoded, decoded, 0, LF);
