+++ src/test/java/org/apache/commons/codec/binary/Base64InputStreamTest.java
@@ -66,7 +66,7 @@ public class Base64InputStreamTest {
 
         // we skip the first character read from the reader
         ins.skip(1);
-        final byte[] decodedBytes = Base64TestData.streamToBytes(ins, new byte[64]);
+        final byte[] decodedBytes = BaseNTestData.streamToBytes(ins, new byte[64]);
         final String str = StringUtils.newStringUtf8(decodedBytes);
 
         assertEquals(STRING_FIXTURE.substring(1), str);
@@ -144,7 +144,7 @@ public class Base64InputStreamTest {
         final Base64InputStream stream = new Base64InputStream(data);
 
         // This line causes an NPE in commons-codec-1.4.jar:
-        final byte[] decodedBytes = Base64TestData.streamToBytes(stream, new byte[1024]);
+        final byte[] decodedBytes = BaseNTestData.streamToBytes(stream, new byte[1024]);
 
         final String decoded = StringUtils.newStringUtf8(decodedBytes);
         assertEquals("codec-98 NPE Base64InputStream", Base64TestData.CODEC_98_NPE_DECODED, decoded);
@@ -219,18 +219,19 @@ public class Base64InputStreamTest {
 
         // OpenSSL interop test.
         encoded = StringUtils.getBytesUtf8(Base64TestData.ENCODED_64_CHARS_PER_LINE);
-        decoded = Base64TestData.DECODED;
+        decoded = BaseNTestData.DECODED;
         testByChunk(encoded, decoded, BaseNCodec.PEM_CHUNK_SIZE, LF);
 
         // Single Line test.
         final String singleLine = Base64TestData.ENCODED_64_CHARS_PER_LINE.replaceAll("\n", "");
         encoded = StringUtils.getBytesUtf8(singleLine);
-        decoded = Base64TestData.DECODED;
+        decoded = BaseNTestData.DECODED;
         testByChunk(encoded, decoded, 0, LF);
 
         // test random data of sizes 0 thru 150
+        final BaseNCodec codec = new Base64(0, null, false);
         for (int i = 0; i <= 150; i++) {
-            final byte[][] randomData = Base64TestData.randomData(i, false);
+            final byte[][] randomData = BaseNTestData.randomData(codec, i);
             encoded = randomData[1];
             decoded = randomData[0];
             testByChunk(encoded, decoded, 0, LF);
@@ -257,18 +258,19 @@ public class Base64InputStreamTest {
 
         // OpenSSL interop test.
         encoded = StringUtils.getBytesUtf8(Base64TestData.ENCODED_64_CHARS_PER_LINE);
-        decoded = Base64TestData.DECODED;
+        decoded = BaseNTestData.DECODED;
         testByteByByte(encoded, decoded, BaseNCodec.PEM_CHUNK_SIZE, LF);
 
         // Single Line test.
         final String singleLine = Base64TestData.ENCODED_64_CHARS_PER_LINE.replaceAll("\n", "");
         encoded = StringUtils.getBytesUtf8(singleLine);
-        decoded = Base64TestData.DECODED;
+        decoded = BaseNTestData.DECODED;
         testByteByByte(encoded, decoded, 0, LF);
 
         // test random data of sizes 0 thru 150
+        final BaseNCodec codec = new Base64(0, null, false);
         for (int i = 0; i <= 150; i++) {
-            final byte[][] randomData = Base64TestData.randomData(i, false);
+            final byte[][] randomData = BaseNTestData.randomData(codec, i);
             encoded = randomData[1];
             decoded = randomData[0];
             testByteByByte(encoded, decoded, 0, LF);
@@ -298,7 +300,7 @@ public class Base64InputStreamTest {
         // Start with encode.
         InputStream in;
         in = new Base64InputStream(new ByteArrayInputStream(decoded), true, chunkSize, separator);
-        byte[] output = Base64TestData.streamToBytes(in);
+        byte[] output = BaseNTestData.streamToBytes(in);
 
         assertEquals("EOF", -1, in.read());
         assertEquals("Still EOF", -1, in.read());
@@ -308,7 +310,7 @@ public class Base64InputStreamTest {
 
         // Now let's try decode.
         in = new Base64InputStream(new ByteArrayInputStream(encoded));
-        output = Base64TestData.streamToBytes(in);
+        output = BaseNTestData.streamToBytes(in);
 
         assertEquals("EOF", -1, in.read());
         assertEquals("Still EOF", -1, in.read());
@@ -320,7 +322,7 @@ public class Base64InputStreamTest {
             in = new Base64InputStream(in, true, chunkSize, separator);
             in = new Base64InputStream(in, false);
         }
-        output = Base64TestData.streamToBytes(in);
+        output = BaseNTestData.streamToBytes(in);
 
         assertEquals("EOF", -1, in.read());
         assertEquals("Still EOF", -1, in.read());
@@ -585,13 +587,13 @@ public class Base64InputStreamTest {
             Base64InputStream in = new Base64InputStream(new ByteArrayInputStream(encoded), false);
             // Default is lenient decoding; it should not throw
             assertFalse(in.isStrictDecoding());
-            Base64TestData.streamToBytes(in);
+            BaseNTestData.streamToBytes(in);
 
             // Strict decoding should throw
             in = new Base64InputStream(new ByteArrayInputStream(encoded), false, 0, null, CodecPolicy.STRICT);
             assertTrue(in.isStrictDecoding());
             try {
-                Base64TestData.streamToBytes(in);
+                BaseNTestData.streamToBytes(in);
                 fail();
             } catch (final IllegalArgumentException ex) {
                 // expected
